Changed the way that new branches are created.

+++ gemla/src/bracket/genetic_state.rs
@@ -7,7 +7,7 @@ pub trait GeneticState {
     /// - iterations: the number of iterations (learning cycles) that the current state should simulate
     /// 
     /// This will be called for every node in a bracket before evaluating it's fitness against other nodes.
-    fn run_simulation(&mut self, iterations: u32);
+    fn run_simulation(&mut self, iterations: u64);
 
     /// Returns a fit score associated with the nodes performance.
     /// This will be used by a bracket in order to determine the most successful child.
